pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "numpy>=2.2.1",
    "numpy-stl>=3.2.0",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.3",
    "werkzeug>=3.1.3",
]
